Tomatoes Score Images
Added logic to determine which RT image to use based on numeric RT
Score. This is not completely accurate for "certified fresh" as RT takes
other factors in consideration beyond the score. Made the ratings
parsing code for robust in case the order in which the ratings sources
are passed changes..